add index just to export stuff
