+++ client/components/accordion/row.tsx
@@ -7,55 +7,25 @@ import React, { forwardRef } from 'react';
 /**
  * Internal dependencies
  */
-import type { AccordionProps, AccordionBodyProps } from './types';
-import AccordionBody from './body';
+import type { AccordionRowProps } from './types';
 import './style.scss';
-
 /**
- * `Accordion` expands and collapses multiple sections of content.
- *
- * ```jsx
- * import { Accordion, AccordionBody, AccordionRow } from '@wordpress/components';
- * import { more } from '@wordpress/icons';
- *
- * const MyAccordion = () => (
- * 	<Accordion header="My Accordion">
- * 		<AccordionBody title="My Block Settings" icon={ more } initialOpen={ true }>
- * 			<AccordionRow>My Accordion Inputs and Labels</AccordionRow>
- * 		</AccordionBody>
- * 	</Accordion>
- * );
- * ```
+ * `AccordionRow` is a generic container for rows within a `AccordionBody`.
+ * It is a flex container with a top margin for spacing.
  */
-const Accordion = forwardRef< HTMLDivElement, AccordionProps >(
-	(
-		{ className, children, highDensity = false, defaultExpanded = false },
-		ref
-	) => {
-		const classNames = clsx( className, 'wcpay-accordion', {
-			'is-high-density': highDensity,
-		} );
-
-		const childrenWithProps = React.Children.map( children, ( child ) => {
-			if (
-				React.isValidElement< AccordionBodyProps >( child ) &&
-				child.type === AccordionBody
-			) {
-				return React.cloneElement( child, {
-					initialOpen: defaultExpanded,
-				} );
-			}
-			return child;
-		} );
-
+const AccordionRow = forwardRef< HTMLDivElement, AccordionRowProps >(
+	( { className, children }, ref ) => {
 		return (
-			<div className={ classNames } ref={ ref }>
-				{ childrenWithProps }
+			<div
+				className={ clsx( 'wcpay-accordion__row', className ) }
+				ref={ ref }
+			>
+				{ children }
 			</div>
 		);
 	}
 );
 
-Accordion.displayName = 'Accordion';
+AccordionRow.displayName = 'AccordionRow';
 
-export default Accordion;
+export default AccordionRow;
